refactor: 컴포넌트 및 스토리 폴더 구조 변경
