+++ src/feature/CardRegister/components/CardPreviewSection/CardExpiryDateDisplay.tsx
