+++ src/feature/CardRegister/components/CardPreviewSection/CardNumberDisplay.tsx
