+++ src/stories/feature/CardRegister/components/InfoInputSection/InfoInputSection.stories.tsx
@@ -2,7 +2,7 @@ import {useState} from 'react';
 import type {Meta, StoryObj} from '@storybook/react-vite';
 import {fn} from 'storybook/test';
 
-import InfoInputSection from './InfoInputSection';
+import InfoInputSection from '../../../../../feature/CardRegister/components/InfoInputSection/InfoInputSection';
 
 const meta = {
   title: 'feature/CardRegister/components/InfoInputSection',
