+++ src/stories/feature/CardRegister/components/InfoInputSection/NumberField.stories.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { fn } from 'storybook/test';
 
-import NumberField from './NumberField';
+import NumberField from '../../../../../feature/CardRegister/components/InfoInputSection/NumberField';
 
 const meta = {
   title: 'feature/CardRegister/components/NumberField',
